Fixing Review comments
